 \name{simRaschData}
 \alias{simRaschData}
 \docType{data}
 \title{Simulated data conforming to the Rasch Model.}
 \description{Dichotomous data simulated from the Rasch model using package eRm.
 }
 \usage{data(simRaschData)}
 \format{Object whit class \code{data.frame} of 300 rows and 15 columns.}
 \keyword{datasets}

 \encoding{latin1}
